+++ scripts/Settings.js
@@ -1,0 +1,20 @@
+export const MODULE_NAME = "Wild Magic Surge 5e";
+export const MODULE_ID = "wild-magic-surge-5e";
+
+export const OPT_ENABLE_CHECK = "enableMagicSurgeCheck";
+export const OPT_CHAT_MSG = "magicSurgeChatMessage";
+export const OPT_AUTO_D20 = "autoRollD20";
+export const OPT_AUTO_D20_MSG = "autoRollD20Message";
+
+export const SPELL_LIST_KEY_WORDS = [
+  "1st Level",
+  "2nd Level",
+  "3rd Level",
+  "4th Level",
+  "5th Level",
+  "6th Level",
+  "7th Level",
+  "8th Level",
+  "9th Level",
+  "10th Level",
+];
